[#313] fix(mypage): totalPostCount 타입을 long으로 변경

+++ src/main/java/com/dduru/gildongmu/post/dto/response/MyPagePostListResponse.java
@@ -14,12 +14,12 @@ public record MyPagePostListResponse(
         @Schema(description = "현재 응답에 포함된 게시글 수", example = "10")
         int size,
         @Schema(description = "삭제되지 않은 내 전체 게시글 수. 전체보기 탭의 '전체 N' 표시에 사용됩니다.", example = "4")
-        int totalPostCount
+        long totalPostCount
 ) {
     public static MyPagePostListResponse of(
             List<MyPagePostSummaryResponse> posts,
             boolean hasNext,
-            int totalPostCount
+            long totalPostCount
     ) {
         Long nextCursor = hasNext && !posts.isEmpty() ? posts.get(posts.size() - 1).id() : null;
         return new MyPagePostListResponse(posts, nextCursor, hasNext, posts.size(), totalPostCount);
